Refactor user transaction
