+++ integration-tests/src/models/account_transaction_models.rs
@@ -13,7 +13,7 @@ use processor::{
     db::{
         models::{
             object_models::v2_object_utils::ObjectWithMetadata,
-            user_transactions_models_temp::user_transactions::UserTransaction,
+            user_transactions_models::user_transactions::UserTransaction,
         },
         resources::FromWriteResource,
     },
