+++ processor/src/db/models/account_transaction_models/account_transactions.rs
@@ -10,7 +10,7 @@ use crate::{
     db::{
         models::{
             object_models::v2_object_utils::ObjectWithMetadata,
-            user_transactions_models_temp::user_transactions::UserTransaction,
+            user_transactions_models::user_transactions::UserTransaction,
         },
         resources::FromWriteResource,
     },
