+++ processor/src/db/models/coin_models/coin_activities.rs
@@ -20,7 +20,7 @@ use crate::{
             v2_fungible_asset_utils::FeeStatement,
         },
         // TODO: update this to use the new signatures after refactoring
-        user_transactions_models_temp::signatures::Signature,
+        user_transactions_models::signatures::Signature,
     },
     schema::coin_activities,
     utils::{
