+++ processor/src/db/models/mod.rs
@@ -8,6 +8,7 @@ pub mod object_models;
 pub mod stake_models;
 pub mod token_models;
 pub mod token_v2_models;
-pub mod user_transactions_models_temp;
+pub mod user_transactions_models;
+
 // Default None value for parquet fields that are not set
 const DEFAULT_NONE: &str = "NULL";
