+++ processor/src/db/models/user_transactions_models/mod.rs
@@ -1,7 +1,6 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod parquet_signatures;
-pub mod parquet_user_transactions;
+// pub mod parquet_signatures;
 pub mod signatures;
 pub mod user_transactions;
